1: useState
React.useState accepts a single argument and returns an array.
The first item is the original state value, and the second item is the
updated state value. As it accepts a new state value it re-renders
the component.

Here we use destructuring to set the first value of useState to name,
and the second to setName. In the handleChange function that is hooked
to the onChange, we update the value of setName to the value our user
is entering. As they enter a value, the setName is updated, and the
component is re-rendered.

If we want to set a prop with a initialName, we can do that. We pass it
to Greeting, and use it in useState as our inital value. Then we add
name as a value to the input, so we can actually see it in the input.
By doing this we are turning value to a controlled input.

We default initialName to an empty string, so that it won't be
undefined, in case no props are passed.