2.0: useEffect:persistent state I got a problem with this one, it doesn't work properlly in chrome, but works just fine in firefox. I haven't been able to figure out why.
Other than that, what we are doing here:
We are using useEffect to run our sideEffect after our event is
trigered and the component re-rendered. In this case we set the value
the user inputs to localStorage.

Web also updated our useState to set the initialName to the value
inside localStorage, if there is one.