2.1 & 2.2 : lazy state initialization & effect dependencies
lazy state initialization
Every time our component functionis run, our function reads from
localStorage. And we only need to know this value the first time this
component is rendered.
We solve this by passing a function instead of a value, it will only
call the function the first time it's rendered.

effect dependencies
useEffect is fired after every re-render. If we want useEffect to fire
only if the source prop has changed instead of after every completed
render, we can pass an array with the values the effect depends on.