2.4: making the custom hook flexible
We make serialize and deserialize so we can put something in
localStorage using JSON.stringify, and pull out using JSON.parse.
This way our value could a number for example, instead of only a string.

We updated our lazy state initialization to work with serialize and
deserialize. And we gave the user the option to pass a function as
the defaultValue.

We maintain the previous key, so if the key changes, we can set the
new key.