3.1: lifting up state
We wanted the display component to display both the name and the
animal. This means we have to keep track of the name state in the
name and display component, and we have to keep track of the animal
state in the animal and display component.

We can do this by lifting the state from name and animal to the
app, their closest common ancestor. This means: we set the state of
both name and animal inside the app, and pass them as props to the
functions that need them. The app owns the shared state, and becomes
the 'source of truth'