3.2 colocate state
'Colocation' is the practice of putting what is relevant to
something close to it. When possible we should keep state close
to it's component. It helps us gain perfomance, and makes our
code easier to maintain.

In this case, since we are no longer displaying the name, the
display function no longer cares about the name state. Now the
name state is only relevant to the name component. So we re-wrote
this code to bring the state into the name component.