4: managed and derived state
We explicitly manage the state of squares, and derive the state of
winner and nextValue from it. The state of status we derive not
only from squares but from the derived states of winner and nextValue.
This way we don't need to worry about updating the derived states
values, they are calculated every render.

In selectSquare we use an early return so that the user won't be
able to change a selected square. And we don't work on our original
array, but on a copy, to avoid bugs. After our copy has been changed
we setSquares to it.

I also wondered why he used .filter(Boolean) inside calculateNextValue,
using it will filter any null or undefined values that might creep into
our array. We are making sure there are only truthy values inside it.